+++ src/components/ChatHistory.js
@@ -1,62 +1,9 @@
-import pdfMake from "pdfmake/build/pdfmake";
-import pdfFonts from "pdfmake/build/vfs_fonts";
-import html2canvas from "html2canvas";
 import React, { memo, useState } from "react";
-import { Box, CircularProgress, TextField, IconButton, Button } from "@mui/material";
+import { Box, CircularProgress, TextField, IconButton } from "@mui/material";
 import ReactMarkdown from "react-markdown";
 import EditIcon from "@mui/icons-material/Edit";
 import { CodeBlock } from "./CodeBlock";
 import { Lightbox } from "react-modal-image";
-
-pdfMake.vfs = pdfFonts.pdfMake.vfs;
-
-const styles = {
-    userMessage: {
-        fontSize: 12,
-        bold: true,
-        alignment: "right",
-        margin: [0, 10, 0, 5],
-    },
-    assistantMessage: {
-        fontSize: 12,
-        margin: [0, 5, 0, 10],
-    },
-};
-
-const generatePdfDocDefinition = async (chatHistory) => {
-    const docDefinition = {
-        content: [],
-        styles: styles,
-    };
-
-    for (const chat of chatHistory) {
-        if (chat.user) {
-            docDefinition.content.push({ text: chat.user, style: "userMessage" });
-        }
-
-        if (chat.assistant) {
-            const content = [];
-            const assistantText = chat.assistant;
-
-            if (chat.image) {
-                const imageDataUrl = Array.isArray(chat.image)
-                    ? await Promise.all(chat.image.map((imgData) => html2canvas(imgData)))
-                    : await html2canvas(chat.image);
-
-                const images = Array.isArray(imageDataUrl)
-                    ? imageDataUrl.map((canvas) => canvas.toDataURL("image/png"))
-                    : [imageDataUrl.toDataURL("image/png")];
-
-                content.push(...images.map((img) => ({ image: img, width: 300 })));
-            }
-
-            content.push({ text: assistantText, style: "assistantMessage" });
-            docDefinition.content.push(...content);
-        }
-    }
-
-    return docDefinition;
-};
 
 const getFileTypeIcon = (mimeType) => {
     switch (mimeType) {
@@ -113,11 +60,6 @@ const ChatHistory = memo(({ chatHistory, isModelResponding, onRun, onChange }) =
     const [editingMessageIndex, setEditingMessageIndex] = useState(-1);
     const [editingMessage, setEditingMessage] = useState("");
 
-    const handleDownloadPDF = async () => {
-        const docDefinition = await generatePdfDocDefinition(chatHistory);
-        pdfMake.createPdf(docDefinition).download("chat_history.pdf");
-    };
-
     const handleImageClick = (index, message) => {
         setLightboxImageIndex(index);
         setLightboxMessageIndex(message);
@@ -146,9 +88,6 @@ const ChatHistory = memo(({ chatHistory, isModelResponding, onRun, onChange }) =
 
     return (
         <Box id="chatid" flex={1} overflow="auto" padding={2} display="flex" flexDirection="column">
-            <Button variant="contained" color="primary" onClick={handleDownloadPDF}>
-                Download Chat History as PDF
-            </Button>
             {chatHistory.map((chat, index) => (
                 <Box
                     data-testid="chat-item"
